I implement the alarm clock widget in a separate file
