+++ py_alarm_clock/__init__.py
@@ -1,0 +1,1 @@
+from .py_alarm_clock import PyAlarmClock
