+++ py_toggle/__init__.py
@@ -1,3 +1,1 @@
 from .py_toggle import PyToggle
-
-
